requirements.txt
Django==4.0.6
msal==1.18.0
celery==5.2.7
redis==4.3.4 # celery broker
django-polymorphic==3.1.0
django-htmx==1.12.2

pre-commit==2.20.0
black==22.6.0
